Break project <> render cycle
Break project <> render cycle

proj
