+++ src/command/check/check.ts
@@ -9,7 +9,9 @@ import { join } from "path/mod.ts";
 import { info } from "log/mod.ts";
 
 import { createSessionTempDir } from "../../core/temp.ts";
-import { render } from "../render/render.ts";
+import { render } from "../render/render-shared.ts";
+import { JupyterCapabilities } from "../../core/jupyter/types.ts";
+import { jupyterCapabilities } from "../../core/jupyter/capabilities.ts";
 import {
   JupyterCapabilities,
   jupyterCapabilities,
