+++ src/command/render/cmd.ts
@@ -12,7 +12,9 @@ import { info } from "log/mod.ts";
 
 import { fixupPandocArgs, kStdOut, parseRenderFlags } from "./flags.ts";
 
-import { render, RenderResult, renderResultFinalOutput } from "./render.ts";
+import { renderResultFinalOutput } from "./render.ts";
+import { render } from "./render-shared.ts";
+import { RenderResult } from "./types.ts";
 
 export const renderCommand = new Command()
   .name("render")
