+++ src/command/render/latexmk/latexmk.ts
@@ -27,7 +27,7 @@ import { Format } from "../../../config/format.ts";
 import { pdfEngine } from "../../../config/pdf.ts";
 
 import { PandocOptions } from "../pandoc.ts";
-import { RenderOptions } from "../render.ts";
+import { RenderOptions } from "../types.ts";
 import { kStdOut, RenderFlags, replacePandocArg } from "../flags.ts";
 import { OutputRecipe } from "../output.ts";
 import { generatePdf } from "./pdf.ts";
