+++ src/command/render/output.ts
@@ -40,7 +40,7 @@ import {
   replacePandocArg,
 } from "./flags.ts";
 import { PandocOptions } from "./pandoc.ts";
-import { RenderContext } from "./render.ts";
+import { RenderContext } from "./types.ts";
 import { resolveKeepSource } from "./codetools.ts";
 
 // render commands imply the --output argument for pandoc and the final
