+++ src/execute/engine.ts
@@ -15,133 +15,19 @@ import {
 } from "../core/yaml.ts";
 import { dirAndStem } from "../core/path.ts";
 
-import { PartitionedMarkdown } from "../core/pandoc/pandoc-partition.ts";
 import { languagesInMarkdown } from "../core/jupyter/jupyter.ts";
 import { restorePreservedHtml } from "../core/jupyter/preserve.ts";
-
-import { ProjectContext } from "../project/project-context.ts";
-
-import { RenderOptions } from "../command/render/render.ts";
-
-import { Format } from "../config/format.ts";
-import { Metadata, metadataAsFormat } from "../config/metadata.ts";
-import {
-  kEngine,
-  kIncludeAfterBody,
-  kIncludeBeforeBody,
-  kIncludeInHeader,
-} from "../config/constants.ts";
+import { metadataAsFormat } from "../config/metadata.ts";
+import { kEngine } from "../config/constants.ts";
 
 import { knitrEngine } from "./rmd.ts";
 import { jupyterEngine } from "./jupyter/jupyter.ts";
 import { markdownEngine } from "./markdown.ts";
-
-export const kQmdExtensions = [".qmd"];
-
-export interface ExecutionEngine {
-  name: string;
-  defaultExt: string;
-  defaultYaml: (kernel?: string) => string[];
-  validExtensions: () => string[];
-  claimsExtension: (ext: string) => boolean;
-  claimsLanguage: (language: string) => boolean;
-  target: (
-    file: string,
-    quiet?: boolean,
-  ) => Promise<ExecutionTarget | undefined>;
-  partitionedMarkdown: (file: string) => Promise<PartitionedMarkdown>;
-  filterFormat?: (
-    source: string,
-    options: RenderOptions,
-    format: Format,
-  ) => Format;
-  execute: (options: ExecuteOptions) => Promise<ExecuteResult>;
-  executeTargetSkipped?: (target: ExecutionTarget, format: Format) => void;
-  dependencies: (options: DependenciesOptions) => Promise<DependenciesResult>;
-  postprocess: (options: PostProcessOptions) => Promise<void>;
-  canFreeze: boolean;
-  canKeepSource?: (target: ExecutionTarget) => boolean;
-  keepFiles?: (input: string) => string[] | undefined;
-  ignoreGlobs?: () => string[] | undefined;
-  devServerRenderOnChange?: (
-    input: string,
-    context: ProjectContext,
-  ) => Promise<boolean>;
-  run?: (options: RunOptions) => Promise<void>;
-}
-
-// execution target (filename and context 'cookie')
-export interface ExecutionTarget {
-  source: string;
-  input: string;
-  metadata: Metadata;
-  data?: unknown;
-}
-
-// execute options
-export interface ExecuteOptions {
-  target: ExecutionTarget;
-  format: Format;
-  resourceDir: string;
-  tempDir: string;
-  dependencies: boolean;
-  libDir?: string;
-  cwd?: string;
-  params?: { [key: string]: unknown };
-  quiet?: boolean;
-}
-
-// result of execution
-export interface ExecuteResult {
-  markdown: string;
-  supporting: string[];
-  filters: string[];
-  includes?: PandocIncludes;
-  engineDependencies?: Array<unknown>;
-  preserve?: Record<string, string>;
-  postProcess?: boolean;
-}
-
-export interface PandocIncludes {
-  [kIncludeBeforeBody]?: string[];
-  [kIncludeAfterBody]?: string[];
-  [kIncludeInHeader]?: string[];
-}
-
-// dependencies options
-export interface DependenciesOptions {
-  target: ExecutionTarget;
-  format: Format;
-  output: string;
-  resourceDir: string;
-  tempDir: string;
-  libDir?: string;
-  dependencies?: Array<unknown>;
-  quiet?: boolean;
-}
-
-// dependencies result
-export interface DependenciesResult {
-  includes: PandocIncludes;
-}
-
-// post processing options
-export interface PostProcessOptions {
-  engine: ExecutionEngine;
-  target: ExecutionTarget;
-  format: Format;
-  output: string;
-  preserve?: Record<string, string>;
-  quiet?: boolean;
-}
-
-// run options
-export interface RunOptions {
-  input: string;
-  render: boolean;
-  port?: number;
-  quiet?: boolean;
-}
+import {
+  ExecutionEngine,
+  ExecutionTarget,
+  PostProcessOptions,
+} from "./types.ts";
 
 const kEngines: ExecutionEngine[] = [
   knitrEngine,
